[IMP] Payment Return Import: display the bank account from the imported file
[UPD] Update account_payment_return_import.pot

Update translation files

Updated by "Update PO files to match POT (msgmerge)" hook in Weblate.

Translation: account-payment-12.0/account-payment-12.0-account_payment_return_import
Translate-URL: https://translation.odoo-community.org/projects/account-payment-12-0/account-payment-12-0-account_payment_return_import/

[UPD] README.rst
