+++ account_payment_return_import/i18n/de.po
@@ -20,13 +20,13 @@ msgstr ""
 "X-Generator: Weblate 3.1.1\n"
 
 #. module: account_payment_return_import
-#: code:addons/account_payment_return_import/wizard/payment_return_import.py:264
+#: code:addons/account_payment_return_import/wizard/payment_return_import.py:267
 #, python-format
 msgid "%d transactions had already been imported and were ignored."
 msgstr "%d Geschäftsvorfälle waren bereits importiert und wurden ignoriert."
 
 #. module: account_payment_return_import
-#: code:addons/account_payment_return_import/wizard/payment_return_import.py:267
+#: code:addons/account_payment_return_import/wizard/payment_return_import.py:270
 #, python-format
 msgid "1 transaction had already been imported and was ignored."
 msgstr "1 Geschäftsvorfall war bereits importiert und wurde ignoriert."
@@ -49,13 +49,24 @@ msgstr ""
 "werden, welches keine automatische Journalerkennung ermöglicht."
 
 #. module: account_payment_return_import
-#: code:addons/account_payment_return_import/wizard/payment_return_import.py:270
+#: code:addons/account_payment_return_import/wizard/payment_return_import.py:273
 #, python-format
 msgid "Already imported items"
 msgstr "Bereits importierte Positionen"
 
 #. module: account_payment_return_import
-#: code:addons/account_payment_return_import/wizard/payment_return_import.py:211
+#: model:ir.model.fields,field_description:account_payment_return_import.field_payment_return__imported_bank_account_id
+#: model_terms:ir.ui.view,arch_db:account_payment_return_import.payment_return_search_view
+msgid "Bank account"
+msgstr ""
+
+#. module: account_payment_return_import
+#: model:ir.model.fields,help:account_payment_return_import.field_payment_return__imported_bank_account_id
+msgid "Bank account from the imported file"
+msgstr ""
+
+#. module: account_payment_return_import
+#: code:addons/account_payment_return_import/wizard/payment_return_import.py:214
 #, python-format
 msgid "Can not determine journal for import."
 msgstr "Kann das Journal für den Import nicht bestimmen."
@@ -178,6 +189,13 @@ msgstr "Nach Import abstimmen"
 #: model:ir.model.fields,field_description:account_payment_return_import.field_payment_return_import__data_file
 msgid "Payment Return File"
 msgstr "Rücklasten-Datei"
+
+#. module: account_payment_return_import
+#: model:ir.model,name:account_payment_return_import.model_payment_return
+#, fuzzy
+#| msgid "Payment return lines"
+msgid "Payment return"
+msgstr "Rückbuchungszeilen"
 
 #. module: account_payment_return_import
 #: model:ir.model,name:account_payment_return_import.model_payment_return_line
