+++ account_payment_return_import/i18n/es.po
@@ -21,13 +21,13 @@ msgstr ""
 "X-Generator: Poedit 1.8.7.1\n"
 
 #. module: account_payment_return_import
-#: code:addons/account_payment_return_import/wizard/payment_return_import.py:264
+#: code:addons/account_payment_return_import/wizard/payment_return_import.py:267
 #, python-format
 msgid "%d transactions had already been imported and were ignored."
 msgstr "%d transacciones ya han sido importadas y serán ignoradas."
 
 #. module: account_payment_return_import
-#: code:addons/account_payment_return_import/wizard/payment_return_import.py:267
+#: code:addons/account_payment_return_import/wizard/payment_return_import.py:270
 #, python-format
 msgid "1 transaction had already been imported and was ignored."
 msgstr "1 transacción ya ha sido importada y será ignorada."
@@ -50,13 +50,24 @@ msgstr ""
 "permiten la detección automática del diario."
 
 #. module: account_payment_return_import
-#: code:addons/account_payment_return_import/wizard/payment_return_import.py:270
+#: code:addons/account_payment_return_import/wizard/payment_return_import.py:273
 #, python-format
 msgid "Already imported items"
 msgstr "Elementos ya importados"
 
 #. module: account_payment_return_import
-#: code:addons/account_payment_return_import/wizard/payment_return_import.py:211
+#: model:ir.model.fields,field_description:account_payment_return_import.field_payment_return__imported_bank_account_id
+#: model_terms:ir.ui.view,arch_db:account_payment_return_import.payment_return_search_view
+msgid "Bank account"
+msgstr ""
+
+#. module: account_payment_return_import
+#: model:ir.model.fields,help:account_payment_return_import.field_payment_return__imported_bank_account_id
+msgid "Bank account from the imported file"
+msgstr ""
+
+#. module: account_payment_return_import
+#: code:addons/account_payment_return_import/wizard/payment_return_import.py:214
 #, python-format
 msgid "Can not determine journal for import."
 msgstr "No puede determinarse el diario para importar."
@@ -180,6 +191,13 @@ msgstr "Casar referencias después de importar"
 #: model:ir.model.fields,field_description:account_payment_return_import.field_payment_return_import__data_file
 msgid "Payment Return File"
 msgstr "Fichero de devoluciones"
+
+#. module: account_payment_return_import
+#: model:ir.model,name:account_payment_return_import.model_payment_return
+#, fuzzy
+#| msgid "Payment return lines"
+msgid "Payment return"
+msgstr "Líneas de la devolución"
 
 #. module: account_payment_return_import
 #: model:ir.model,name:account_payment_return_import.model_payment_return_line
